requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-telegram-bot==21.4
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
